+++ lbp-simulator/vec4.h
@@ -8,14 +8,14 @@
 /** X, Y, Z, U in micrometers. */
 struct Vec4
 {
-	int32_t x = 0;
-	int32_t y = 0;
-	int32_t z = 0;
-	int32_t u = 0;
+	int64_t x = 0;
+	int64_t y = 0;
+	int64_t z = 0;
+	int64_t u = 0;
 
 	/** Default constructor. All Zeros. */
 	Vec4() = default;
-	Vec4(int32_t x, int32_t y, int32_t z, int32_t u);
+	Vec4(int64_t x, int64_t y, int64_t z, int64_t u);
 
 	/** @return True if any of the four values are non-zero. */
 	explicit operator bool() const;
@@ -26,7 +26,7 @@ Vec4 operator+(const Vec4 &a, const Vec4 &b);
 Vec4 operator-(const Vec4 &a, const Vec4 &b);
 bool operator==(const Vec4 &a, const Vec4 &b);
 bool operator!=(const Vec4 &a, const Vec4 &b);
-Vec4 operator*(const Vec4 &vec, int32_t m);
-Vec4 operator/(const Vec4 &vec, int32_t d);
+Vec4 operator*(const Vec4 &vec, int64_t m);
+Vec4 operator/(const Vec4 &vec, int64_t d);
 
 Vec4 clamp(const Vec4 &vec, const Vec4 &limit);
